python_flask/requirements.txt
Flask
gunicorn
flask-cors
pyjwt